LRIS-33072 fix user segments

+++ modules/dxp/apps/osb/osb-customer/osb-customer-zendesk-documentation-sync/osb-customer-zendesk-documentation-sync-service/src/main/java/com/liferay/osb/customer/zendesk/documentation/sync/configuration/ZendeskDocumentationSyncConfigurationValues.java
@@ -21,8 +21,8 @@ import com.liferay.portal.kernel.util.GetterUtil;
  */
 public class ZendeskDocumentationSyncConfigurationValues {
 
-	public static final int ZENDESK_SECTION_USER_SEGMENT_ID =
-		GetterUtil.getInteger(
+	public static final long ZENDESK_SECTION_USER_SEGMENT_ID =
+		GetterUtil.getLong(
 			ZendeskDocumentationSyncConfigurationUtil.get(
 				"zendesk.section.user.segment.id"));
